Develop be

+++ server/photomory/src/main/java/com/example/photomory/repository/MyAlbumRepository.java
@@ -3,5 +3,5 @@ package com.example.photomory.repository;
 import com.example.photomory.entity.MyAlbum;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MyAlbumRepository extends JpaRepository<MyAlbum, Long> {
+public interface MyAlbumRepository extends JpaRepository<MyAlbum, Integer> {
 }
